Removed dictionaries from AllTypes as this class is used to test insert/createFromJson methods which are not supported for dictionaries at the moment
