+++ realm/realm-library/src/testUtils/java/io/realm/entities/AllTypes.java
@@ -25,7 +25,6 @@ import java.util.UUID;
 
 import io.realm.Mixed;
 import io.realm.MutableRealmInteger;
-import io.realm.RealmDictionary;
 import io.realm.RealmList;
 import io.realm.RealmObject;
 import io.realm.RealmResults;
@@ -61,49 +60,10 @@ public class AllTypes extends RealmObject {
     public static final String FIELD_FLOAT_LIST = "columnFloatList";
     public static final String FIELD_DATE_LIST = "columnDateList";
 
-    public static final String FIELD_BOOLEAN_DICTIONARY = "columnBooleanDictionary";
-    public static final String FIELD_STRING_DICTIONARY = "columnStringDictionary";
-    public static final String FIELD_INTEGER_DICTIONARY = "columnIntegerDictionary";
-    public static final String FIELD_FLOAT_DICTIONARY = "columnFloatDictionary";
-    public static final String FIELD_LONG_DICTIONARY = "columnLongDictionary";
-    public static final String FIELD_SHORT_DICTIONARY = "columnShortDictionary";
-    public static final String FIELD_DOUBLE_DICTIONARY = "columnDoubleDictionary";
-    public static final String FIELD_BYTE_DICTIONARY = "columnByteDictionary";
-    public static final String FIELD_BINARY_DICTIONARY = "columnBinaryDictionary";
-    public static final String FIELD_DATE_DICTIONARY = "columnDateDictionary";
-    public static final String FIELD_OBJECT_ID_DICTIONARY = "columnObjectIdDictionary";
-    public static final String FIELD_UUID_DICTIONARY = "columnUUIDDictionary";
-    public static final String FIELD_DECIMAL128_DICTIONARY = "columnDecimal128Dictionary";
-    public static final String FIELD_MIXED_DICTIONARY = "columnMixedDictionary";
-
-    public static final String FIELD_REQUIRED_BOOLEAN_DICTIONARY = "columnRequiredBooleanDictionary";
-    public static final String FIELD_REQUIRED_STRING_DICTIONARY = "columnRequiredStringDictionary";
-    public static final String FIELD_REQUIRED_INTEGER_DICTIONARY = "columnRequiredIntegerDictionary";
-    public static final String FIELD_REQUIRED_FLOAT_DICTIONARY = "columnRequiredFloatDictionary";
-    public static final String FIELD_REQUIRED_LONG_DICTIONARY = "columnRequiredLongDictionary";
-    public static final String FIELD_REQUIRED_SHORT_DICTIONARY = "columnRequiredShortDictionary";
-    public static final String FIELD_REQUIRED_DOUBLE_DICTIONARY = "columnRequiredDoubleDictionary";
-    public static final String FIELD_REQUIRED_BYTE_DICTIONARY = "columnRequiredByteDictionary";
-    public static final String FIELD_REQUIRED_BINARY_DICTIONARY = "columnRequiredBinaryDictionary";
-    public static final String FIELD_REQUIRED_BOXED_BINARY_DICTIONARY = "columnRequiredBoxedBinaryDictionary";
-    public static final String FIELD_REQUIRED_DATE_DICTIONARY = "columnRequiredDateDictionary";
-    public static final String FIELD_REQUIRED_OBJECT_ID_DICTIONARY = "columnRequiredObjectIdDictionary";
-    public static final String FIELD_REQUIRED_UUID_DICTIONARY = "columnRequiredUUIDDictionary";
-    public static final String FIELD_REQUIRED_DECIMAL128_DICTIONARY = "columnRequiredDecimal128Dictionary";
-
     public static final String[] INVALID_TYPES_FIELDS_FOR_DISTINCT
             = new String[]{FIELD_REALMOBJECT, FIELD_REALMLIST, FIELD_DOUBLE, FIELD_FLOAT,
             FIELD_STRING_LIST, FIELD_BINARY_LIST, FIELD_BOOLEAN_LIST, FIELD_LONG_LIST,
-            FIELD_DOUBLE_LIST, FIELD_FLOAT_LIST, FIELD_DATE_LIST, FIELD_BOOLEAN_DICTIONARY,
-            FIELD_STRING_DICTIONARY, FIELD_INTEGER_DICTIONARY, FIELD_FLOAT_DICTIONARY,
-            FIELD_LONG_DICTIONARY, FIELD_SHORT_DICTIONARY, FIELD_DOUBLE_DICTIONARY,
-            FIELD_BYTE_DICTIONARY, FIELD_BINARY_DICTIONARY,
-            FIELD_DATE_DICTIONARY, FIELD_OBJECT_ID_DICTIONARY, FIELD_UUID_DICTIONARY,
-            FIELD_DECIMAL128_DICTIONARY, FIELD_MIXED_DICTIONARY, FIELD_REQUIRED_BOOLEAN_DICTIONARY,
-            FIELD_REQUIRED_STRING_DICTIONARY, FIELD_REQUIRED_INTEGER_DICTIONARY, FIELD_REQUIRED_FLOAT_DICTIONARY,
-            FIELD_REQUIRED_LONG_DICTIONARY, FIELD_REQUIRED_SHORT_DICTIONARY, FIELD_REQUIRED_DOUBLE_DICTIONARY,
-            FIELD_REQUIRED_BYTE_DICTIONARY, FIELD_REQUIRED_BINARY_DICTIONARY, FIELD_REQUIRED_BOXED_BINARY_DICTIONARY,
-            FIELD_REQUIRED_DATE_DICTIONARY, FIELD_REQUIRED_OBJECT_ID_DICTIONARY, FIELD_REQUIRED_UUID_DICTIONARY};
+            FIELD_DOUBLE_LIST, FIELD_FLOAT_LIST, FIELD_DATE_LIST};
 
 
     @Required
@@ -146,50 +106,6 @@ public class AllTypes extends RealmObject {
     private RealmList<ObjectId> columnObjectIdList;
     private RealmList<UUID> columnUUIDList;
     private RealmList<Mixed> columnMixedList;
-
-    private RealmDictionary<DogPrimaryKey> columnRealmDictionary;
-
-    private RealmDictionary<Boolean> columnBooleanDictionary;
-    private RealmDictionary<String> columnStringDictionary;
-    private RealmDictionary<Integer> columnIntegerDictionary;
-    private RealmDictionary<Float> columnFloatDictionary;
-    private RealmDictionary<Long> columnLongDictionary;
-    private RealmDictionary<Short> columnShortDictionary;
-    private RealmDictionary<Double> columnDoubleDictionary;
-    private RealmDictionary<Byte> columnByteDictionary;
-    private RealmDictionary<byte[]> columnBinaryDictionary;
-    private RealmDictionary<Date> columnDateDictionary;
-    private RealmDictionary<ObjectId> columnObjectIdDictionary;
-    private RealmDictionary<UUID> columnUUIDDictionary;
-    private RealmDictionary<Decimal128> columnDecimal128Dictionary;
-    private RealmDictionary<Mixed> columnMixedDictionary;
-
-    @Required
-    private RealmDictionary<Boolean> columnRequiredBooleanDictionary;
-    @Required
-    private RealmDictionary<String> columnRequiredStringDictionary;
-    @Required
-    private RealmDictionary<Integer> columnRequiredIntegerDictionary;
-    @Required
-    private RealmDictionary<Float> columnRequiredFloatDictionary;
-    @Required
-    private RealmDictionary<Long> columnRequiredLongDictionary;
-    @Required
-    private RealmDictionary<Short> columnRequiredShortDictionary;
-    @Required
-    private RealmDictionary<Double> columnRequiredDoubleDictionary;
-    @Required
-    private RealmDictionary<Byte> columnRequiredByteDictionary;
-    @Required
-    private RealmDictionary<byte[]> columnRequiredBinaryDictionary;
-    @Required
-    private RealmDictionary<Date> columnRequiredDateDictionary;
-    @Required
-    private RealmDictionary<ObjectId> columnRequiredObjectIdDictionary;
-    @Required
-    private RealmDictionary<UUID> columnRequiredUUIDDictionary;
-    @Required
-    private RealmDictionary<Decimal128> columnRequiredDecimal128Dictionary;
 
     public String getColumnString() {
         return columnString;
@@ -402,231 +318,5 @@ public class AllTypes extends RealmObject {
 
     public void setColumnMixedList(RealmList<Mixed> columnMixedList) {
         this.columnMixedList = columnMixedList;
-    }
-
-
-    public RealmDictionary<DogPrimaryKey> getColumnRealmDictionary() {
-        return columnRealmDictionary;
-    }
-
-    public void setColumnRealmDictionary(RealmDictionary<DogPrimaryKey> columnRealmDictionary) {
-        this.columnRealmDictionary = columnRealmDictionary;
-    }
-
-    public RealmDictionary<Boolean> getColumnBooleanDictionary() {
-        return columnBooleanDictionary;
-    }
-
-    public void setColumnBooleanDictionary(RealmDictionary<Boolean> columnBooleanDictionary) {
-        this.columnBooleanDictionary = columnBooleanDictionary;
-    }
-
-    public RealmDictionary<String> getColumnStringDictionary() {
-        return columnStringDictionary;
-    }
-
-    public void setColumnStringDictionary(RealmDictionary<String> columnStringDictionary) {
-        this.columnStringDictionary = columnStringDictionary;
-    }
-
-    public RealmDictionary<Integer> getColumnIntegerDictionary() {
-        return columnIntegerDictionary;
-    }
-
-    public void setColumnIntegerDictionary(RealmDictionary<Integer> columnIntegerDictionary) {
-        this.columnIntegerDictionary = columnIntegerDictionary;
-    }
-
-    public RealmDictionary<Float> getColumnFloatDictionary() {
-        return columnFloatDictionary;
-    }
-
-    public void setColumnFloatDictionary(RealmDictionary<Float> columnFloatDictionary) {
-        this.columnFloatDictionary = columnFloatDictionary;
-    }
-
-    public RealmDictionary<Long> getColumnLongDictionary() {
-        return columnLongDictionary;
-    }
-
-    public void setColumnLongDictionary(RealmDictionary<Long> columnLongDictionary) {
-        this.columnLongDictionary = columnLongDictionary;
-    }
-
-    public RealmDictionary<Short> getColumnShortDictionary() {
-        return columnShortDictionary;
-    }
-
-    public void setColumnShortDictionary(RealmDictionary<Short> columnShortDictionary) {
-        this.columnShortDictionary = columnShortDictionary;
-    }
-
-    public RealmDictionary<Double> getColumnDoubleDictionary() {
-        return columnDoubleDictionary;
-    }
-
-    public void setColumnDoubleDictionary(RealmDictionary<Double> columnDoubleDictionary) {
-        this.columnDoubleDictionary = columnDoubleDictionary;
-    }
-
-    public RealmDictionary<Byte> getColumnByteDictionary() {
-        return columnByteDictionary;
-    }
-
-    public void setColumnByteDictionary(RealmDictionary<Byte> columnByteDictionary) {
-        this.columnByteDictionary = columnByteDictionary;
-    }
-
-    public RealmDictionary<byte[]> getColumnBinaryDictionary() {
-        return columnBinaryDictionary;
-    }
-
-    public void setColumnBinaryDictionary(RealmDictionary<byte[]> columnBinaryDictionary) {
-        this.columnBinaryDictionary = columnBinaryDictionary;
-    }
-
-    public RealmDictionary<Date> getColumnDateDictionary() {
-        return columnDateDictionary;
-    }
-
-    public void setColumnDateDictionary(RealmDictionary<Date> columnDateDictionary) {
-        this.columnDateDictionary = columnDateDictionary;
-    }
-
-    public RealmDictionary<ObjectId> getColumnObjectIdDictionary() {
-        return columnObjectIdDictionary;
-    }
-
-    public void setColumnObjectIdDictionary(RealmDictionary<ObjectId> columnObjectIdDictionary) {
-        this.columnObjectIdDictionary = columnObjectIdDictionary;
-    }
-
-    public RealmDictionary<UUID> getColumnUUIDDictionary() {
-        return columnUUIDDictionary;
-    }
-
-    public void setColumnUUIDDictionary(RealmDictionary<UUID> columnUUIDDictionary) {
-        this.columnUUIDDictionary = columnUUIDDictionary;
-    }
-
-    public RealmDictionary<Decimal128> getColumnDecimal128Dictionary() {
-        return columnDecimal128Dictionary;
-    }
-
-    public void setColumnDecimal128Dictionary(RealmDictionary<Decimal128> columnDecimal128Dictionary) {
-        this.columnDecimal128Dictionary = columnDecimal128Dictionary;
-    }
-
-    public RealmDictionary<Mixed> getColumnMixedDictionary() {
-        return columnMixedDictionary;
-    }
-
-    public void setColumnMixedDictionary(RealmDictionary<Mixed> columnMixedDictionary) {
-        this.columnMixedDictionary = columnMixedDictionary;
-    }
-
-    public RealmDictionary<Boolean> getColumnRequiredBooleanDictionary() {
-        return columnRequiredBooleanDictionary;
-    }
-
-    public void setColumnRequiredBooleanDictionary(RealmDictionary<Boolean> columnRequiredBooleanDictionary) {
-        this.columnRequiredBooleanDictionary = columnRequiredBooleanDictionary;
-    }
-
-    public RealmDictionary<String> getColumnRequiredStringDictionary() {
-        return columnRequiredStringDictionary;
-    }
-
-    public void setColumnRequiredStringDictionary(RealmDictionary<String> columnRequiredStringDictionary) {
-        this.columnRequiredStringDictionary = columnRequiredStringDictionary;
-    }
-
-    public RealmDictionary<Integer> getColumnRequiredIntegerDictionary() {
-        return columnRequiredIntegerDictionary;
-    }
-
-    public void setColumnRequiredIntegerDictionary(RealmDictionary<Integer> columnRequiredIntegerDictionary) {
-        this.columnRequiredIntegerDictionary = columnRequiredIntegerDictionary;
-    }
-
-    public RealmDictionary<Float> getColumnRequiredFloatDictionary() {
-        return columnRequiredFloatDictionary;
-    }
-
-    public void setColumnRequiredFloatDictionary(RealmDictionary<Float> columnRequiredFloatDictionary) {
-        this.columnRequiredFloatDictionary = columnRequiredFloatDictionary;
-    }
-
-    public RealmDictionary<Long> getColumnRequiredLongDictionary() {
-        return columnRequiredLongDictionary;
-    }
-
-    public void setColumnRequiredLongDictionary(RealmDictionary<Long> columnRequiredLongDictionary) {
-        this.columnRequiredLongDictionary = columnRequiredLongDictionary;
-    }
-
-    public RealmDictionary<Short> getColumnRequiredShortDictionary() {
-        return columnRequiredShortDictionary;
-    }
-
-    public void setColumnRequiredShortDictionary(RealmDictionary<Short> columnRequiredShortDictionary) {
-        this.columnRequiredShortDictionary = columnRequiredShortDictionary;
-    }
-
-    public RealmDictionary<Double> getColumnRequiredDoubleDictionary() {
-        return columnRequiredDoubleDictionary;
-    }
-
-    public void setColumnRequiredDoubleDictionary(RealmDictionary<Double> columnRequiredDoubleDictionary) {
-        this.columnRequiredDoubleDictionary = columnRequiredDoubleDictionary;
-    }
-
-    public RealmDictionary<Byte> getColumnRequiredByteDictionary() {
-        return columnRequiredByteDictionary;
-    }
-
-    public void setColumnRequiredByteDictionary(RealmDictionary<Byte> columnRequiredByteDictionary) {
-        this.columnRequiredByteDictionary = columnRequiredByteDictionary;
-    }
-
-    public RealmDictionary<byte[]> getColumnRequiredBinaryDictionary() {
-        return columnRequiredBinaryDictionary;
-    }
-
-    public void setColumnRequiredBinaryDictionary(RealmDictionary<byte[]> columnRequiredBinaryDictionary) {
-        this.columnRequiredBinaryDictionary = columnRequiredBinaryDictionary;
-    }
-
-
-    public RealmDictionary<Date> getColumnRequiredDateDictionary() {
-        return columnRequiredDateDictionary;
-    }
-
-    public void setColumnRequiredDateDictionary(RealmDictionary<Date> columnRequiredDateDictionary) {
-        this.columnRequiredDateDictionary = columnRequiredDateDictionary;
-    }
-
-    public RealmDictionary<ObjectId> getColumnRequiredObjectIdDictionary() {
-        return columnRequiredObjectIdDictionary;
-    }
-
-    public void setColumnRequiredObjectIdDictionary(RealmDictionary<ObjectId> columnRequiredObjectIdDictionary) {
-        this.columnRequiredObjectIdDictionary = columnRequiredObjectIdDictionary;
-    }
-
-    public RealmDictionary<UUID> getColumnRequiredUUIDDictionary() {
-        return columnRequiredUUIDDictionary;
-    }
-
-    public void setColumnRequiredUUIDDictionary(RealmDictionary<UUID> columnRequiredUUIDDictionary) {
-        this.columnRequiredUUIDDictionary = columnRequiredUUIDDictionary;
-    }
-
-    public RealmDictionary<Decimal128> getColumnRequiredDecimal128Dictionary() {
-        return columnRequiredDecimal128Dictionary;
-    }
-
-    public void setColumnRequiredDecimal128Dictionary(RealmDictionary<Decimal128> columnRequiredDecimal128Dictionary) {
-        this.columnRequiredDecimal128Dictionary = columnRequiredDecimal128Dictionary;
     }
 }
